backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aioredis==2.0.1
alembic==1.13.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
